requirements.txt
alabaster==0.7.16
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
<<<<<<< HEAD
=======
<<<<<<< HEAD
=======
dj-rest-auth==7.0.1
>>>>>>> 82b70b0 (Fixes)
>>>>>>> 3ce4b0e (FIxes)
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-filter==24.3
django-phonenumber-field==8.0.0
django-rest-auth==0.9.5
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
<<<<<<< HEAD
djangorestframework-simplejwt==5.3.1
=======
<<<<<<< HEAD
djangorestframework-simplejwt==5.3.1
=======
djangorestframework_simplejwt==5.5.0
>>>>>>> 82b70b0 (Fixes)
>>>>>>> 3ce4b0e (FIxes)
docutils==0.20.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
<<<<<<< HEAD
jwt==1.3.1
=======
<<<<<<< HEAD
jwt==1.3.1
=======
>>>>>>> 82b70b0 (Fixes)
>>>>>>> 3ce4b0e (FIxes)
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.20
numpy==2.2.4
packaging==24.1
phonenumbers==8.13.46
pillow==10.4.0
pipdeptree==2.16.2
pkginfo==1.12.0
platformdirs==4.3.6
psutil==6.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pyfiglet==1.0.2
Pygments==2.18.0
<<<<<<< HEAD
PyJWT==2.1.0
=======
<<<<<<< HEAD
PyJWT==2.1.0
=======
PyJWT==2.9.0
>>>>>>> 82b70b0 (Fixes)
>>>>>>> 3ce4b0e (FIxes)
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rfc3986==2.0.0
rpds-py==0.20.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tabulate==0.9.0
termcolor==2.5.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==14.1
wheel==0.45.1
whitenoise==6.8.2
<<<<<<< HEAD
wrapt==1.17.0
=======
<<<<<<< HEAD
wrapt==1.17.0
=======
wrapt==1.17.0
>>>>>>> 82b70b0 (Fixes)
>>>>>>> 3ce4b0e (FIxes)
